requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.3.0
backoff==2.2.1
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
dataclasses-json==0.6.7
defusedxml==0.7.1
Django==5.2
filelock==3.18.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.3.2
huggingface-hub==0.30.1
idna==3.10
inference-sdk==0.45.1
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
mpmath==1.3.0
multidict==6.3.1
mypy-extensions==1.0.0
networkx==3.4.2
numpy==1.26.4
opencv-python==4.10.0.84
packaging==24.2
pillow==10.4.0
propcache==0.3.1
protobuf==6.30.2
py-cpuinfo==9.0.0
pyparsing==3.2.3
pytesseract==0.3.13
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scipy==1.15.2
sentencepiece==0.2.0
setuptools==78.1.0
six==1.17.0
sqlparse==0.5.3
supervision==0.25.1
sympy==1.13.1
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
transformers==4.50.3
typing-inspect==0.9.0
typing_extensions==4.13.0
urllib3==2.3.0
yarl==1.18.3
